updatew

+++ be/src/vec/columns/column_decimal.cpp
@@ -348,7 +348,7 @@ ColumnPtr ColumnDecimal<T>::replicate(const IColumn::Offsets& offsets) const {
 
 template <typename T>
 void ColumnDecimal<T>::replicate(const uint32_t* counts, size_t target_size, IColumn& column,
-                                 size_t begin, size_t count_sz) const {
+                                 size_t begin, int count_sz) const {
     size_t size = count_sz < 0 ? data.size() : count_sz;
     if (0 == size) return;
